rama_ai/requirements.txt
# RAMA AI v2.0 - Requirements
# Fully offline AI Voice Assistant

# Core
asyncio>=3.4.3
pydantic>=2.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0

# AI/ML - Local
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0

# Voice - Offline STT/TTS
# Install one of:
# openai-whisper (larger, more accurate)
# faster-whisper (faster, less accurate)
# For 8GB RAM: use faster-whisper with tiny model
faster-whisper>=0.10.0

# TTS - Use one of:
# coqui-tts (good quality)
# piper-tts (fast, low RAM)
# silero (good quality, easy)
silero>=1.0.0

# Wake Word
pvporcupine>=2.1.0

# Ingestion
pypdf>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
chromadb>=0.4.0

# UI
customtkinter>=5.2.0
matplotlib>=3.7.0

# Utilities
requests>=2.28.0
aiohttp>=3.8.0
plyer>=2.1.0

# For code parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0